requirements.txt
africastalking==1.2.7
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==4.34
colorama==0.4.6
contextlib2==21.6.0
cryptography==41.0.5
dataclasses-json==0.6.2
distro==1.8.0
fastapi==0.104.1
fastavro==1.8.2
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
halo==0.0.31
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.334
langsmith==0.0.63
log-symbols==0.0.14
MarkupSafe==2.1.3
marshmallow==3.20.1
MouseInfo==0.1.3
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pypdf==3.17.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.0
python-multipart==0.0.6
pytweening==1.0.7
PyYAML==6.0.1
requests==2.31.0
schema==0.7.5
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spinners==0.0.24
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
termcolor==2.3.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
validators==0.22.0
watchfiles==0.21.0
weaviate-client==3.25.3
websockets==12.0
Werkzeug==3.0.1
wikipedia==1.4.0
yarl==1.9.2
zipp==3.17.0
